requirements.txt
torch==2.0.1+cu117
torchvision==0.15.2+cu117
torchaudio==2.0.2+cu117
tensorboard
numpy==1.26.4
opencv-python
jupyter
matplotlib
pandas
pillow
scipy
tqdm
h5py
ipykernel
torch-summary
fiftyone
nuscenes-devkit
open3d
